move account size to deletion screen and align all icons/text in hover info
